+++ react/utils/messages.ts
@@ -30,6 +30,7 @@ export const messages = defineMessages({
   withoutStock: { id: 'store/checkout.b2b.withoutStock' },
   emptyCart: { id: 'store/checkout.b2b.empty.cart' },
   clearCart: { id: 'store/checkout.b2b.button.clearCart' },
+  removeItem: { id: 'store/checkout.b2b.button.removeItem' },
   searchPlaceholder: { id: 'store/checkout.b2b.search.placeholder' },
   placeOrder: { id: 'store/checkout.b2b.button.placeOrder' },
 })